refactor: move FeeConsensus into modules
