+++ fedimintd/Cargo.toml
@@ -35,6 +35,7 @@ fedimint-lnv2-server = { workspace = true }
 fedimint-logging = { workspace = true }
 fedimint-meta-server = { workspace = true }
 fedimint-metrics = { workspace = true }
+fedimint-mint-common = { workspace = true }
 fedimint-mint-server = { workspace = true }
 fedimint-rocksdb = { workspace = true }
 fedimint-server = { workspace = true }
