+++ modules/fedimint-lnv2-common/src/config.rs
@@ -4,8 +4,7 @@ pub use bitcoin::Network;
 use fedimint_core::core::ModuleKind;
 use fedimint_core::encoding::{Decodable, Encodable};
 use fedimint_core::envs::BitcoinRpcConfig;
-use fedimint_core::fee_consensus::FeeConsensus;
-use fedimint_core::{plugin_types_trait_impl_config, PeerId};
+use fedimint_core::{plugin_types_trait_impl_config, Amount, PeerId};
 use group::Curve;
 use serde::{Deserialize, Serialize};
 use tpe::{AggregatePublicKey, PublicKeyShare, SecretKeyShare};
@@ -24,7 +23,7 @@ impl LightningGenParams {
         Self {
             local: LightningGenParamsLocal { bitcoin_rpc },
             consensus: LightningGenParamsConsensus {
-                fee_consensus: FeeConsensus::new_lnv2(1000).expect("Relative fee is within range"),
+                fee_consensus: FeeConsensus::new(1000).expect("Relative fee is within range"),
                 network: Network::Regtest,
             },
         }
@@ -94,6 +93,78 @@ plugin_types_trait_impl_config!(
     LightningClientConfig
 );
 
+#[derive(Debug, Clone, Eq, PartialEq, Hash, Serialize, Deserialize, Encodable, Decodable)]
+pub struct FeeConsensus {
+    base: Amount,
+    parts_per_million: u64,
+}
+
+impl FeeConsensus {
+    /// The lightning module will charge a non-configurable base fee of one
+    /// satoshi per transaction input and output to account for the costs
+    /// incurred by the federation for processing the transaction. On top of
+    /// that the federation may charge a additional relative fee per input and
+    /// output of up to one thousand parts per million which is equal to one
+    /// tenth of one percent.
+    ///
+    /// # Errors
+    /// - This constructor returns an error if the relative fee is in excess of
+    ///   one thousand parts per million.
+    pub fn new(parts_per_million: u64) -> anyhow::Result<Self> {
+        anyhow::ensure!(
+            parts_per_million <= 1_000,
+            "Relative fee over one thousand parts per million is excessive"
+        );
+
+        Ok(Self {
+            base: Amount::from_sats(1),
+            parts_per_million,
+        })
+    }
+
+    pub fn fee(&self, amount: Amount) -> Amount {
+        Amount::from_msats(self.fee_msats(amount.msats))
+    }
+
+    fn fee_msats(&self, msats: u64) -> u64 {
+        msats
+            .saturating_mul(self.parts_per_million)
+            .saturating_div(1_000_000)
+            .checked_add(self.base.msats)
+            .expect("The division creates sufficient headroom to add the base fee")
+    }
+}
+
+#[test]
+fn test_fee_consensus() {
+    let fee_consensus = FeeConsensus::new(1_000).expect("Relative fee is within range");
+
+    assert_eq!(
+        fee_consensus.fee(Amount::from_msats(999)),
+        Amount::from_sats(1)
+    );
+
+    assert_eq!(
+        fee_consensus.fee(Amount::from_sats(1)),
+        Amount::from_msats(1) + Amount::from_sats(1)
+    );
+
+    assert_eq!(
+        fee_consensus.fee(Amount::from_sats(1000)),
+        Amount::from_sats(1) + Amount::from_sats(1)
+    );
+
+    assert_eq!(
+        fee_consensus.fee(Amount::from_bitcoins(1)),
+        Amount::from_sats(100_000) + Amount::from_sats(1)
+    );
+
+    assert_eq!(
+        fee_consensus.fee(Amount::from_bitcoins(100_000)),
+        Amount::from_bitcoins(100) + Amount::from_sats(1)
+    );
+}
+
 #[allow(dead_code)]
 fn migrate_config_consensus(
     config: &fedimint_ln_common::config::LightningConfigConsensus,
@@ -116,7 +187,7 @@ fn migrate_config_consensus(
                 )
             })
             .collect(),
-        fee_consensus: FeeConsensus::new_lnv2(1000).expect("Relative fee is within range"),
+        fee_consensus: FeeConsensus::new(1000).expect("Relative fee is within range"),
         network: config.network,
     }
 }
