+++ modules/fedimint-mint-tests/tests/tests.rs
@@ -7,7 +7,6 @@ use fedimint_client::transaction::{ClientInput, ClientInputBundle, TransactionBu
 use fedimint_core::bitcoin_migration::bitcoin30_to_bitcoin32_keypair;
 use fedimint_core::config::EmptyGenParams;
 use fedimint_core::core::OperationId;
-use fedimint_core::fee_consensus::FeeConsensus;
 use fedimint_core::secp256k1::KeyPair;
 use fedimint_core::task::sleep_in_test;
 use fedimint_core::util::NextOrPending;
@@ -20,7 +19,7 @@ use fedimint_mint_client::{
     MintClientInit, MintClientModule, Note, OOBNotes, ReissueExternalNotesState,
     SelectNotesWithAtleastAmount, SpendOOBState,
 };
-use fedimint_mint_common::config::{MintGenParams, MintGenParamsConsensus};
+use fedimint_mint_common::config::{FeeConsensus, MintGenParams, MintGenParamsConsensus};
 use fedimint_mint_common::{MintInput, MintInputV0, Nonce};
 use fedimint_mint_server::MintInit;
 use fedimint_testing::fixtures::{Fixtures, TIMEOUT};
